pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mental-health-chatbot"
version = "0.1.0"
description = "HIPAA-compliant AI chatbot for mental health therapists using GoHighLevel"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    # Core AI Framework
    "pydantic-ai>=0.0.14",
    "dspy-ai>=2.5.0",
    "instructor>=1.0.0",
    "marvin>=3.0.0",
    
    # GoHighLevel MCP Integration
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "websockets>=12.0",
    
    # HIPAA Compliance & Security
    "cryptography>=42.0.0",
    "pycryptodome>=3.20.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    
    # Database & Caching (HIPAA-compliant)
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "redis[hiredis]>=5.0.0",
    
    # Web Framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "starlette>=0.36.0",
    
    # Mental Health & AI Safety
    "langchain-community>=0.0.20",
    "transformers>=4.40.0",
    "torch>=2.2.0",
    
    # Utilities
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.2",
    "typer>=0.9.0",
    "rich>=13.7.0",
    
    # Testing & Development
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.6.0",
    "pytest-mock>=3.12.0",
    "factory-boy>=3.3.0",
    "freezegun>=1.4.0",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "hipaa: marks tests that verify HIPAA compliance",
]